requirements.txt
pyannote.audio==3.1.1
transformers==4.30.2
huggingface_hub
python-dotenv
pydub
PySide6
torch==2.7.1
torchaudio==2.7.1
noisereduce
scipy
